Simplify useToggle test file
